app/requirements/local.txt
aiohttp==3.7.4.post0
apipkg==1.5
apparmor==3.0.3
appdirs==1.4.4
asgiref==3.4.1
asn1crypto==1.4.0
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
Automat==20.2.0
Babel==2.9.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
boto3==1.18.33
botocore==1.21.33
bs4==0.0.1
cached-property==1.5.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cmdln==2.0.0
colorama==0.4.4
commonmark==0.9.1
configobj==5.0.6
constantly==15.1.0
crit==0.0.1
cryptography==3.4.7
cssselect==1.1.0
cupshelpers==1.0
cycler==0.10.0
dateutils==0.6.12
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.1
distro==1.6.0
Django==3.2.5
django-appconf==1.0.4
django-bootstrap4==3.0.1
django-countries==7.2.1
django-debug-toolbar==3.2.2
django-encrypted-fields==1.1.2
django-environ==0.5.0
django-extensions==3.1.3
django-filter==2.4.0
django-ipware==3.0.2
django-tracking-analyzer==1.1.1
django-user-agents==0.4.0
djangorestframework==3.12.4
dnspython==2.1.0
docker==4.4.4
docker-compose==1.28.5
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
ecdsa==0.16.1
entrypoints==0.3
filelock==3.0.12
fluidity-sm==0.2.0
geoip2==4.2.0
h2==4.0.0
hpack==4.0.0
html5lib==1.1
httplib2==0.19.1
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.2
importlib-metadata==3.7.2
incremental==21.3.0
iniconfig==0.0.0
invoke==1.6.0
ipaddr==2.2.0
ipykernel==5.4.3
ipyparallel==6.3.0
ipython==7.25.0
ipython_genutils==0.2.0
ipywidgets==7.6.3
isc==2.0
isodate==0.6.0
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.1
kiwisolver==1.3.1
lexicon==1.0.0
Markdown==3.3.4
matplotlib-inline==0.1.2
maxminddb==2.0.3
mistune==0.8.4
more-itertools==8.8.0
multidict==5.1.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
nftables==0.1
notebook==6.2.0
notify2==0.3.1
ntlm-auth==1.5.0
numpy==1.21.0
olefile==0.46
ordered-set==4.0.2
outcome==1.0.1
packaging==20.9
pandas==1.3.2
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.8.2
passlib==1.7.4
pbkdf2==1.3
pbr==5.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.0
pip-api==0.0.14
pip-search==0.0.7
pip-tools==5.4.0
plotly==5.1.0
pluggy==0.13.1
ply==3.11
prometheus-client==0.11.0
prompt-toolkit==3.0.8
protobuf==3.17.3
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pycrypto==2.6.1
Pygments==2.9.0
PyHamcrest==2.0.2
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.18.0
pyserial==3.5
PySocks==1.7.1
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.15.0
python-keyczar==0.716
python-linux-procfs==0.6
python-memcached==1.58
pytz==2021.1
pyudev==0.22.0
pyvmomi==7.0.2
pywinrm==0.4.1
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.1.0
QtPy==1.10.0
rdflib==5.0.0
redis==3.5.3
requests==2.25.1
requests-ntlm==1.1.0
requests-toolbelt==0.9.1
rich==10.6.0
s3transfer==0.5.0
Send2Trash==1.5.0
service-identity==18.1.0
sh==1.14.2
simplegeneric==0.8.1
six==1.16.0
slip==0.6.5
slip.dbus==0.6.5
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.1
stevedore==3.3.0
tenacity==8.0.0
terminado==0.9.5
testpath==0.4.4
texttable==1.6.3
toml==0.10.2
tornado==6.1
traitlets==4.3.3
trio==0.17.0
Twisted==21.7.0
typing-extensions==3.10.0.0
ua-parser==0.10.0
urllib3==1.26.6
user-agents==2.2.0
virtualenv==20.2.2
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
xmltodict==0.12.0
yarl==1.6.3
zipp==3.5.0
zope.event==4.5.0
zope.interface==5.4.0
